app/requirements.txt
dash==2.13.0
dash-bootstrap-components==1.4.1
dash-leaflet==0.1.21
geopandas==0.13.0
pandas==1.5.3
plotly==5.14.1
SQLAlchemy==1.4.46
python-dotenv==1.0.0
flask-caching==2.3.0
numpy==1.21.6
pytest==7.2.0
psycopg2-binary==2.9.6
Flask==2.2.5
Flask-Login==0.6.3
Flask-Security-Too==4.0.0
Werkzeug==2.2.3
